+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- A `version` command to show the current version of `hoc`.
+
 ## [0.0.2] - 2022-11-30
 
 ### Added